requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
commonmark==0.9.1
fastapi==0.88.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
plotly==5.13.1
psutil==5.9.4
pydantic==1.10.2
Pygments==2.14.0
pynecone==0.1.20
pynecone-io==0.1.14
python-dotenv==1.0.0
python-engineio==4.4.0
python-multipart==0.0.5
python-socketio==5.8.0
redis==4.5.3
requests==2.28.2
rfc3986==1.5.0
rich==12.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.22.0
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typer==0.4.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
watchdog==2.3.1
websockets==10.4
yarl==1.8.2
